refactor(manager/fleet): rename types.ts
