+++ lib/modules/manager/fleet/schema.ts
